move classes to folder
