+++ src/main/java/com/auth/jwtmicroservice/auth/AuthenticationController.java
@@ -1,6 +1,6 @@
 package com.auth.jwtmicroservice.auth;
 
-import com.auth.jwtmicroservice.config.FrontendConfigProperties;
+import com.auth.jwtmicroservice.config.ConfigProperties.FrontendConfigProperties;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
